requirements.txt
streamlit
pandas
docxtpl
python-docx
requests
openpyxl
numpy
docx2pdf
